+++ Isotop2/Data/Entities/ProgramSettings.cs
@@ -3,13 +3,13 @@
     public class ProgramSettings
     {
         public string FileName = Environment.CurrentDirectory + "\\Program_Settings.xml";
-        public decimal NewGenerationActivity = 19000M;
-        public decimal OldGenerationActivity = 19000M;
+        public double NewGenerationActivity = 19000;
+        public double OldGenerationActivity = 19000;
         public int TimeDecay = 0;
         public DateTime DateOnZeroDay = new DateTime(2025, 1, 1);
-        public decimal IodineActivity = 200M;
-        public decimal RadiumActivity = 10.2M;
+        public double IodineActivity = 200;
+        public double RadiumActivity = 10.2;
         public DateTime CreateDateRadium = new DateTime(2025, 1, 1);
-        public decimal PatientWeighet = 70M;
+        public int PatientWeighet = 70;
     }
 }
